Move Air Conditioning prerequisite from buildings to research project
Because buildings depended on multiple things, some mods get a bit confused and display confusing information to end-users, leading them to believe that either research project was all that was needed. Even in the base game, this was somewhat present since the research tree did not make it clear that Air Conditioning is a prerequisite.
This change moves the dependency upon the Air Conditioning research project from the buildings to the Basic Climate Control research project. This both fixes the research tree to explicitly require/show the dependency and prevents mods from being confused about which project is actually needed.
Resolves upstream #14